Add files via upload
Project is about "Sustainable Supply Chain Performance Dashboard" by using Power Bi : It is about creating a comprehensive Dashboard for managing and analyzing supply chain inventory data, manufacturing data and supplier data. This is my work of the first week of my AICTE internship in which i have transformed the given dataset as per the requirements.